+++ src/tests/Mocks/Mocks.csproj
@@ -3,7 +3,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>TestMocks</RootNamespace>
+    <RootNamespace>Mocks</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
