+++ src/tests/Tests.Linux/Tests.Linux.csproj
@@ -6,7 +6,7 @@
 
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
-    <RootNamespace>LinuxTests</RootNamespace>
+    <RootNamespace>Tests.Linux</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
