+++ src/tests/Tests.Windows/FileIOTests.cs
@@ -1,7 +1,4 @@
-using FileSurfer.Windows.Models.FileInformation;
-using FileSurfer.Windows.Services.FileOperations;
-
-namespace FSTests;
+namespace Tests.Windows;
 
 public class FileIoTests
 {
