+++ src/tests/Tests.Windows/Tests.Windows.csproj
@@ -6,7 +6,7 @@
 
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
-    <RootNamespace>WindowsTests</RootNamespace>
+    <RootNamespace>Tests.Windows</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
